github/requirements.txt
gspread
oauth2client
google-generativeai
schedule
python-dotenv
